requirements.txt
aiohttp
chromadb
cryptography
faiss-cpu
fastapi
fastmcp
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
groq
langchain
langchain-community
langchain-core
langchain-groq
langchain_openai
langgraph
loguru
mcp-tavily
mcp-use
passlib[bcrypt]
pymupdf
pypdf
python-dotenv
python-multipart
requests
sentence-transformers
sqlalchemy
tavily-python
typesense
uv
uvicorn
itsdangerous